refactor(http): move RequestManager to http
